+++ NEWS.md
@@ -11,6 +11,10 @@
     - Saves motif plots as PDFs. 
 * Added `get_bpparam` function.
 
+## Bug fixes
+
+* `get_SNPgroup_counts`: Get rid of `.dots` and `.groups` warnings.
+
 # echoannot 0.99.8
 
 ## New features